Restructure CPU and PPU to share one Memory object
Due to Rust’s ownership policy, the Memory object in main.rs can’t be passed to the CPU and PPU during both of their initializations. The code has been refactored to pass the Memory object to the CPU and PPU when running their respective step functions. This allows the CPU to use the memory as needed first, then return ownership so the PPU can do the same.